july14 updateded class

+++ challenge4.java
@@ -4,7 +4,7 @@ public class challenge4 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int num1 = sc.nextInt();
+        long num1 = sc.nextLong();
 
         if (num1 % 2 == 0)
             System.out.println("This number is even.");
